LPS-159174 Move to individual classes - no logic changes
