+++ modules/apps/client-extension/client-extension-web/src/main/java/com/liferay/client/extension/web/internal/upgrade/registry/ClientExtensionWebUpgradeStepRegistrator.java
@@ -14,20 +14,9 @@
 
 package com.liferay.client.extension.web.internal.upgrade.registry;
 
-import com.liferay.portal.kernel.log.Log;
-import com.liferay.portal.kernel.log.LogFactoryUtil;
 import com.liferay.portal.kernel.model.Release;
-import com.liferay.portal.kernel.upgrade.BasePortletIdUpgradeProcess;
 import com.liferay.portal.kernel.upgrade.UpgradeProcess;
-import com.liferay.portal.kernel.util.ArrayUtil;
 import com.liferay.portal.upgrade.registry.UpgradeStepRegistrator;
-
-import java.sql.Connection;
-import java.sql.ResultSet;
-import java.sql.Statement;
-
-import java.util.ArrayList;
-import java.util.List;
 
 import org.osgi.service.component.annotations.Component;
 import org.osgi.service.component.annotations.Reference;
@@ -44,17 +33,8 @@ public class ClientExtensionWebUpgradeStepRegistrator
 	public void register(Registry registry) {
 		registry.register(
 			"0.0.0", "1.0.0",
-			new BasePortletIdUpgradeProcess() {
-
-				@Override
-				protected String[][] getRenamePortletIdsArray() {
-					return _getRenamePortletIdsArray(
-						connection, "remote_app_",
-						"com_liferay_remote_app_web_internal_portlet_" +
-							"RemoteAppEntryPortlet_");
-				}
-
-			});
+			new com.liferay.client.extension.web.internal.upgrade.v1_0_0.
+				UpgradePortletId());
 
 		registry.register(
 			"1.0.0", "2.0.0",
@@ -68,60 +48,9 @@ public class ClientExtensionWebUpgradeStepRegistrator
 				}
 
 			},
-			new BasePortletIdUpgradeProcess() {
-
-				@Override
-				protected String[][] getRenamePortletIdsArray() {
-					return ArrayUtil.append(
-						new String[][] {
-							{
-								"com_liferay_remote_app_admin_web_portlet_" +
-									"RemoteAppAdminPortlet",
-								"com_liferay_client_extension_web_internal_" +
-									"portlet_ClientExtensionAdminPortlet"
-							}
-						},
-						_getRenamePortletIdsArray(
-							connection,
-							"com_liferay_remote_app_web_internal_portlet_" +
-								"RemoteAppEntryPortlet_",
-							"com_liferay_client_extension_web_internal_" +
-								"portlet_ClientExtensionEntryPortlet_"));
-				}
-
-			});
+			new com.liferay.client.extension.web.internal.upgrade.v2_0_0.
+				UpgradePortletId());
 	}
-
-	private String[][] _getRenamePortletIdsArray(
-		Connection connection, String oldPortletIdPrefix,
-		String newPortletIdPrefix) {
-
-		List<String[]> portletIds = new ArrayList<>();
-
-		try (Statement statement = connection.createStatement();
-			ResultSet resultSet = statement.executeQuery(
-				"select clientExtensionEntryId from ClientExtensionEntry ")) {
-
-			while (resultSet.next()) {
-				long clientExtensionEntryId = resultSet.getLong(
-					"clientExtensionEntryId");
-
-				portletIds.add(
-					new String[] {
-						oldPortletIdPrefix + clientExtensionEntryId,
-						newPortletIdPrefix + clientExtensionEntryId
-					});
-			}
-		}
-		catch (Exception exception) {
-			_log.error(exception);
-		}
-
-		return portletIds.toArray(new String[0][0]);
-	}
-
-	private static final Log _log = LogFactoryUtil.getLog(
-		ClientExtensionWebUpgradeStepRegistrator.class);
 
 	@Reference(
 		target = "(&(release.bundle.symbolic.name=com.liferay.client.extension.service)(release.schema.version>=3.0.0))"
